fix(sheet): remove unused commands,remove set range values unused param
