+++ packages/base-sheets/src/commands/mutations/set-range-values.mutation.ts
@@ -1,12 +1,16 @@
 import {
     CellValueType,
     CommandType,
+    IBorderData,
     ICellData,
     ICellV,
     ICopyToOptionsData,
+    IDocumentData,
+    IKeyValue,
     IMutation,
     IRange,
     IStyleData,
+    ITextStyle,
     IUniverInstanceService,
     Nullable,
     ObjectMatrix,
@@ -15,11 +19,8 @@ import {
 } from '@univerjs/core';
 import { IAccessor } from '@wendellhu/redi';
 
-import { mergeRichTextStyle, mergeStyle, transformStyle } from './set-border-styles.mutation';
-
 /** Params of `SetRangeValuesMutation` */
 export interface ISetRangeValuesMutationParams {
-    range: IRange[]; // FIXME: maybe don't need this
     worksheetId: string;
     workbookId: string;
 
@@ -32,6 +33,10 @@ export interface ISetRangeValuesMutationParams {
      * @deprecated not a good design
      */
     options?: ICopyToOptionsData;
+}
+
+export interface ISetRangeValuesRangeMutationParams extends ISetRangeValuesMutationParams {
+    range: IRange[];
 }
 
 /**
@@ -149,7 +154,6 @@ export const SetRangeValuesMutation: IMutation<ISetRangeValuesMutationParams, bo
                 if (newVal.m !== undefined) {
                     oldVal.m = newVal.m;
                 }
-                console.info('oldVaue======', oldVal);
 
                 if (newVal.t !== undefined) {
                     oldVal.t = newVal.t;
@@ -210,4 +214,144 @@ function checkCellValueType(v: Nullable<ICellV>): Nullable<CellValueType> {
         return CellValueType.BOOLEAN;
     }
     return CellValueType.FORCE_STRING;
+}
+
+/**
+ * Convert old style data for storage
+ * @param style
+ */
+export function transformStyle(oldStyle: Nullable<IStyleData>, newStyle: Nullable<IStyleData>): Nullable<IStyleData> {
+    const backupStyle = transformNormalKey(oldStyle, newStyle);
+    return backupStyle;
+}
+/**
+ * Convert old style normal key for storage
+ *
+ * @param style
+ */
+export function transformNormalKey(
+    oldStyle: Nullable<IStyleData>,
+    newStyle: Nullable<IStyleData>
+): Nullable<IStyleData> {
+    // If there is no newly set style, directly store the historical style
+    if (!newStyle || !Object.keys(newStyle).length) {
+        return oldStyle;
+    }
+    const backupStyle: { [key: string]: any } = Tools.deepClone(oldStyle) || {};
+
+    for (const k in newStyle) {
+        if (k === 'bd') {
+            backupStyle[k] = transformBorders(backupStyle[k] || {}, newStyle[k]);
+        }
+        // 1. To modify the existing style,we need original setting to undo
+        // 2. Newly set the style, we need null to undo
+        else if (!(k in backupStyle)) {
+            backupStyle[k] = null;
+        }
+    }
+    return backupStyle;
+}
+/**
+ * Convert old style border for storage
+ *
+
+ * @param style
+ */
+export function transformBorders(oldBorders: IBorderData, newBorders: Nullable<IBorderData>): IBorderData {
+    // If there is no newly set border, directly store the historical border
+    if (!newBorders || !Object.keys(newBorders).length) {
+        return oldBorders;
+    }
+
+    for (const k in newBorders) {
+        // 1. To modify the existing border,we need original setting to undo
+        // 2. Newly set the border, we need null to undo
+        if (!(k in oldBorders)) {
+            (oldBorders as IKeyValue)[k] = null;
+        }
+    }
+
+    return oldBorders;
+}
+
+/**
+ * merge new style to old style
+ *
+ * @param oldStyle
+ * @param newStyle
+ */
+export function mergeStyle(
+    oldStyle: Nullable<IStyleData>,
+    newStyle: Nullable<IStyleData>,
+    isRichText: boolean = false
+): Nullable<IStyleData> {
+    // clear style
+    if (newStyle === null) return newStyle;
+    // don't operate
+    if (newStyle === undefined) return oldStyle;
+
+    const backupStyle: { [key: string]: any } = Tools.deepClone(oldStyle) || {};
+    if (!backupStyle) return;
+    for (const k in newStyle) {
+        if (isRichText && ['bd', 'tr', 'td', 'ht', 'vt', 'tb', 'pd'].includes(k)) {
+            continue;
+        }
+        // you can only choose one of the themeColor and rgbColor of the border setting
+        if (k in backupStyle && k === 'bd') {
+            backupStyle[k] = Object.assign(backupStyle[k], newStyle[k]);
+        } else {
+            backupStyle[k] = (newStyle as IKeyValue)[k];
+        }
+    }
+
+    // Overline/Strikethrough/Underline color follows text color
+    // if ('cl' in backupStyle) {
+    //     for (const k in newStyle) {
+    //         if (['ul', 'ol', 'st'].includes(k)) {
+    //             backupStyle[k].cl = backupStyle.cl;
+    //         }
+    //     }
+    // }
+
+    if ('cl' in backupStyle) {
+        if ('ul' in backupStyle) {
+            backupStyle.ul.cl = backupStyle.cl;
+        }
+
+        if ('ol' in backupStyle) {
+            backupStyle.ol.cl = backupStyle.cl;
+        }
+
+        if ('st' in backupStyle) {
+            backupStyle.st.cl = backupStyle.cl;
+        }
+    }
+
+    return backupStyle;
+}
+
+/**
+ * Find the text style of all paragraphs and modify it to the new style
+ * @param p
+ * @param newStyle
+ */
+export function mergeRichTextStyle(p: IDocumentData, newStyle: Nullable<IStyleData>) {
+    p.body?.textRuns?.forEach((textRun) => {
+        if (!textRun.ts) {
+            textRun.ts = {};
+        }
+
+        const oldStyle = textRun.ts || {};
+
+        const merge = mergeStyle(oldStyle as Nullable<IStyleData>, newStyle, true);
+
+        // then remove null
+        merge && Tools.removeNull(merge);
+
+        if (Tools.isEmptyObject(merge)) {
+            delete textRun.ts;
+        } else {
+            textRun.ts = merge as ITextStyle;
+        }
+    });
 }
